// src/types/AccommodationType.d.ts
export type AccommodationDetailInfoType = {
  roomimg4alt: "string";
  roomimg5: "string";
  roomimg5alt: "string";
  contentid: "string";
  contenttypeid: "string";
  fldgubun: "string";
  infoname: "string";
  infotext: "string";
  serialnum: "string";
  subcontentid: "string";
  subdetailalt: "string";
  subdetailimg: "string";
  subdetailoverview: "string";
  subname: "string";
  subnum: "string";
  roomcode: "string";
  roomtitle: "string";
  roomsize1: "string";
  roomcount: "string";
  roombasecount: "string";
  roommaxcount: "string";
  roomoffseasonminfee1: "string";
  roomoffseasonminfee2: "string";
  roompeakseasonminfee1: "string";
  roompeakseasonminfee2: "string";
  roomintro: "string";
  roombathfacility: "string";
  roombath: "string";
  roomhometheater: "string";
  roomaircondition: "string";
  roomtv: "string";
  roompc: "string";
  roomcable: "string";
  roominternet: "string";
  roomrefrigerator: "string";
  roomtoiletries: "string";
  roomsofa: "string";
  roomcook: "string";
  roomtable: "string";
  roomhairdryer: "string";
  roomsize2: "string";
  roomimg1: "string";
  roomimg1alt: "string";
  roomimg2: "string";
  roomimg2alt: "string";
  roomimg3: "string";
  roomimg3alt: "string";
  roomimg4: "string";
};

export type AccommodationDetailIntroType = {
  opendate: "string";
  parking: "string";
  restdate: "string";
  useseason: "string";
  usetime: "string";
  accomcountculture: "string";
  chkbabycarriageculture: "string";
  chkcreditcardculture: "string";
  chkpetculture: "string";
  discountinfo: "string";
  infocenterculture: "string";
  parkingculture: "string";
  parkingfee: "string";
  restdateculture: "string";
  usefee: "string";
  usetimeculture: "string";
  scale: "string";
  spendtime: "string";
  agelimit: "string";
  bookingplace: "string";
  usetimeleports: "string";
  accomcountlodging: "string";
  benikia: "string";
  checkintime: "string";
  checkouttime: "string";
  chkcooking: "string";
  foodplace: "string";
  goodstay: "string";
  hanok: "string";
  infocenterlodging: "string";
  parkinglodging: "string";
  pickup: "string";
  roomcount: "string";
  reservationlodging: "string";
  reservationurl: "string";
  roomtype: "string";
  scalelodging: "string";
  subfacility: "string";
  barbecue: "string";
  beauty: "string";
  beverage: "string";
  bicycle: "string";
  campfire: "string";
  fitness: "string";
  karaoke: "string";
  publicbath: "string";
  eventenddate: "string";
  eventhomepage: "string";
  eventplace: "string";
  eventstartdate: "string";
  festivalgrade: "string";
  placeinfo: "string";
  playtime: "string";
  program: "string";
  spendtimefestival: "string";
  sponsor1: "string";
  sponsor1tel: "string";
  sponsor2: "string";
  sponsor2tel: "string";
  subevent: "string";
  usetimefestival: "string";
  distance: "string";
  infocentertourcourse: "string";
  schedule: "string";
  publicpc: "string";
  sauna: "string";
  seminar: "string";
  sports: "string";
  refundregulation: "string";
  chkbabycarriageshopping: "string";
  chkcreditcardshopping: "string";
  chkpetshopping: "string";
  culturecenter: "string";
  fairday: "string";
  infocentershopping: "string";
  opendateshopping: "string";
  opentime: "string";
  parkingshopping: "string";
  restdateshopping: "string";
  restroom: "string";
  saleitem: "string";
  saleitemcost: "string";
  scaleshopping: "string";
  shopguide: "string";
  chkcreditcardfood: "string";
  discountinfofood: "string";
  firstmenu: "string";
  infocenterfood: "string";
  kidsfacility: "string";
  opendatefood: "string";
  opentimefood: "string";
  packing: "string";
  parkingfood: "string";
  reservationfood: "string";
  restdatefood: "string";
  scalefood: "string";
  seat: "string";
  smoking: "string";
  treatmenu: "string";
  lcnsno: "string";
  contentid: "string";
  contenttypeid: "string";
  accomcount: "string";
  chkbabycarriage: "string";
  chkcreditcard: "string";
  chkpet: "string";
  expagerange: "string";
  expguide: "string";
  heritage1: "string";
  heritage2: "string";
  heritage3: "string";
  infocenter: "string";
  taketime: "string";
  theme: "string";
  accomcountleports: "string";
  chkbabycarriageleports: "string";
  chkcreditcardleports: "string";
  chkpetleports: "string";
  expagerangeleports: "string";
  infocenterleports: "string";
  openperiod: "string";
  parkingfeeleports: "string";
  parkingleports: "string";
  reservation: "string";
  restdateleports: "string";
  scaleleports: "string";
  usefeeleports: "string";
  discountinfofestival: "string";
};